+++ Contentstack.Management.Core/Models/ContentModelling.cs
@@ -1,55 +1,53 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
+using System.Text.Json.Serialization;
 using Contentstack.Management.Core.Models.Fields;
-using Newtonsoft.Json;
 
 namespace Contentstack.Management.Core.Models
 {
-    [JsonObject(ItemNullValueHandling = NullValueHandling.Ignore)]
     public class ContentModelling
     {
-        [JsonProperty(propertyName: "title")]
-        public string Title { get; set; }
+        [JsonPropertyName("title")]
+        public string? Title { get; set; }
 
-        [JsonProperty(propertyName: "uid")]
-        public string Uid { get; set; }
+        [JsonPropertyName("uid")]
+        public string? Uid { get; set; }
 
-        [JsonProperty(propertyName: "description")]
-        public string Description { get; set; }
+        [JsonPropertyName("description")]
+        public string? Description { get; set; }
 
-        [JsonProperty(propertyName: "field_rules")]
-        public List<FieldRules> FieldRules { get; set; }
+        [JsonPropertyName("field_rules")]
+        public List<FieldRules>? FieldRules { get; set; }
 
-        [JsonProperty(propertyName: "schema")]
-        public List<Field> Schema { get; set; }
+        [JsonPropertyName("schema")]
+        public List<Field>? Schema { get; set; }
 
-        [JsonProperty(propertyName: "global_field_refs")]
-        public List<GlobalFieldRefs> GlobalFieldRefs { get; set; }
+        [JsonPropertyName("global_field_refs")]
+        public List<GlobalFieldRefs>? GlobalFieldRefs { get; set; }
 
-        [JsonProperty(propertyName: "options")]
-        public Option Options { get; set; }
+        [JsonPropertyName("options")]
+        public Option? Options { get; set; }
     }
 
-    [JsonObject(ItemNullValueHandling = NullValueHandling.Ignore)]
     public class Option
     {
-        [JsonProperty(propertyName: "title")]
-        public string Title { get; set; }
+        [JsonPropertyName("title")]
+        public string? Title { get; set; }
 
-        [JsonProperty(propertyName: "sub_title")]
-        public List<string> SubTitle { get; set; }
+        [JsonPropertyName("sub_title")]
+        public List<string>? SubTitle { get; set; }
 
-        [JsonProperty(propertyName: "singleton")]
+        [JsonPropertyName("singleton")]
         public bool Singleton { get; set; }
 
-        [JsonProperty(propertyName: "is_page")]
+        [JsonPropertyName("is_page")]
         public bool IsPage { get; set; }
 
-        [JsonProperty(propertyName: "url_pattern")]
-        public string UrlPattern { get; set; }
+        [JsonPropertyName("url_pattern")]
+        public string? UrlPattern { get; set; }
 
-        [JsonProperty(propertyName: "url_prefix")]
-        public string UrlPrefix { get; set; }
+        [JsonPropertyName("url_prefix")]
+        public string? UrlPrefix { get; set; }
 
     }    
 }