+++ Contentstack.Management.Core/Models/Fields/ExtensionField.cs
@@ -1,14 +1,15 @@
 ﻿using System;
 using System.Collections.Generic;
-using Newtonsoft.Json;
+using System.Text.Json;
+using System.Text.Json.Serialization;
 
 namespace Contentstack.Management.Core.Models.Fields
 {
     public class ExtensionField : Field
     {
-        [JsonProperty(propertyName: "extension_uid")]
+        [JsonPropertyName("extension_uid")]
         public string extension_uid { get; set; }
-        [JsonProperty(propertyName: "config")]
-        public Dictionary<string, object> config { get; set; }
+        [JsonPropertyName("config")]
+        public Dictionary<string, JsonElement> config { get; set; }
     }
 }