+++ Contentstack.Management.Core/Models/Fields/Field.cs
@@ -1,45 +1,44 @@
 using System;
-using Newtonsoft.Json;
+using System.Text.Json.Serialization;
 
 namespace Contentstack.Management.Core.Models.Fields
 {
-    [JsonObject(ItemNullValueHandling = NullValueHandling.Ignore)]
     public class Field
     {
         /// <summary>
         /// Determines the display name of a field. It is a mandatory field.
         /// </summary>
-        [JsonProperty(propertyName: "display_name")]
-        public string DisplayName { get; set; }
+        [JsonPropertyName("display_name")]
+        public string? DisplayName { get; set; }
         /// <summary>
         /// Represents the unique ID of each field. It is a mandatory field.	
         /// </summary>
-        [JsonProperty(propertyName: "uid")]
-        public string Uid { get; set; }
+        [JsonPropertyName("uid")]
+        public string? Uid { get; set; }
         /// <summary>
         /// Determines what value can be provided to the Title field.	
         /// </summary>
-        [JsonProperty(propertyName: "data_type")]
-        public string DataType { get; set; }
+        [JsonPropertyName("data_type")]
+        public string? DataType { get; set; }
         /// <summary>
         /// Allows you to enter additional data about a field. Also, you can add additional values under 'field_metadata'.
         /// </summary>
-        [JsonProperty(propertyName: "field_metadata")]
-        public FieldMetadata FieldMetadata { get; set; }
+        [JsonPropertyName("field_metadata")]
+        public FieldMetadata? FieldMetadata { get; set; }
 
-        [JsonProperty(propertyName: "multiple")]
+        [JsonPropertyName("multiple")]
         public bool Multiple { get; set; }
 
-        [JsonProperty(propertyName: "mandatory")]
+        [JsonPropertyName("mandatory")]
         public bool Mandatory { get; set; }
 
-        [JsonProperty(propertyName: "unique")]
+        [JsonPropertyName("unique")]
         public bool Unique { get; set; }
 
         /// <summary>
         /// Presentation widget for text fields (e.g. dropdown, checkbox).
         /// </summary>
-        [JsonProperty(propertyName: "display_type")]
-        public string DisplayType { get; set; }
+        [JsonPropertyName("display_type")]
+        public string? DisplayType { get; set; }
     }
 }