+++ Contentstack.Management.Core/Models/Fields/JsonField.cs
@@ -1,4 +1,5 @@
-using Newtonsoft.Json;
+using System.Text.Json;
+using System.Text.Json.Serialization;
 
 namespace Contentstack.Management.Core.Models.Fields
 {
@@ -7,7 +8,7 @@ namespace Contentstack.Management.Core.Models.Fields
     /// </summary>
     public class JsonField : TextboxField
     {
-        [JsonProperty(propertyName: "reference_to")]
-        public object ReferenceTo { get; set; }
+        [JsonPropertyName("reference_to")]
+        public JsonElement? ReferenceTo { get; set; }
     }
 }